comment changes #7

+++ src/main/kotlin/io/openfuture/state/blockchain/Blockchain.kt
@@ -3,9 +3,9 @@ package io.openfuture.state.blockchain
 import io.openfuture.state.blockchain.dto.UnifiedBlock
 
 abstract class Blockchain {
-    abstract suspend fun getLastBlockNumber(): Long
+    abstract suspend fun getLastBlockNumber(): Int
 
-    abstract suspend fun getBlock(blockNumber: Long): UnifiedBlock
+    abstract suspend fun getBlock(blockNumber: Int): UnifiedBlock
 
     fun getName(): String = javaClass.simpleName
